Move Database\JoinClause to Database\Query\Join
